+++ lib/src/config/global_config.dart
@@ -1,0 +1,7 @@
+import '../../rx_shared_preferences.dart';
+
+/// Global config for default singleton or extension.
+class RxSharedPreferencesConfig {
+  /// Config for logger.
+  static Logger logger = const DefaultLogger();
+}